freeze migrations into v6.5.2_1.ts, bump to 6.5.2:2
